refactor: move DEBUG code paths from the main SceneDelegate to a new DebuggignSceneDelegate subclass to separate debug and test-specific code from production code
