+++ EssentialApp/EssentialApp/DebuggingSceneDelegate.swift
@@ -1,0 +1,46 @@
+//
+//  SceneDelegate.swift
+//  EssentialApp
+//
+//  Created by Nikhil Menon on 5/23/24.
+//
+
+import UIKit
+import EssentialFeed
+import CoreData
+
+#if DEBUG
+class DebuggingSceneDelegate: SceneDelegate {
+    
+    override func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+        
+        if CommandLine.arguments.contains("-reset") {
+            try? FileManager.default.removeItem(at: localStoreURL)
+        }
+        
+        super.scene(scene, willConnectTo: session, options: connectionOptions)
+    }
+    
+    override func makeRemoteClient() -> HTTPClient {
+        if UserDefaults.standard.string(forKey: "connectivity") == "offline" {
+            return AlwaysFailingHTTPClient()
+        }
+        
+        return super.makeRemoteClient()
+    }
+}
+
+fileprivate class AlwaysFailingHTTPClient: HTTPClient {
+    
+    private class Task: HTTPClientTask {
+        func cancel() { }
+    }
+    
+    func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> any EssentialFeed.HTTPClientTask {
+        let error = NSError(domain: "offline", code: 0)
+        completion(.failure(error))
+        return Task()
+    }
+}
+#endif
